// src/app/shared/model/cust-company-obj.model.ts
export class CustCompanyObj {
  CustCompanyId: number;
  CustId: number;
  MrCompanyTypeCode: string;
  RefIndustryTypeId: number;
  RegistrationNo: string;
  LicenseNo: string;
  NumOfEmp: Number;
  MrInvestmentTypeCode: string;
  EstablishmentDt: Date;
  IsAffiliated: boolean;
  Website: string;
  PhnArea1: string;
  Phn1: string;
  PhnExt1: string;
  PhnArea2: string;
  Phn2: string;
  PhnExt2: string;
  Email1: string;
  Email2: string;
  IsSkt: boolean;
  IsVip: boolean;
  VipNotes: string;
  MrCustModelCode: string;
  IsAffiliateWithMf: boolean;
  RefSectorEconomySlikId: number;
  ParentCustId: number;
  RowVersion: string;
}